requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pandas
numpy
Pillow
opencv-python-headless
requests
earthengine-api
